backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
PyYAML==6.0.2
Jinja2==3.1.4
python-multipart==0.0.9
click==8.1.7
rich==13.7.1
orjson==3.10.6
aiofiles==24.1.0
loguru==0.7.2

